preliminary added parser, storage, ui and tasklist classes
